Add webpack bundle building to node-wot-td-tools
The `node-wot-td-tools` package is useful not only in server scenarios
using `node`, but also in browsers. To support usage in browsers, the
classes need to be bundled, resolving the modules used in the sources.

This change introduces `webpack` to create a bundle file as part of the
build process. You will find the resulting bundle in
`dist/node-wot-td-tools.bundle.js`.

To use classes are available on the global `NodeWotTdTools` object.